+++ Platipus.Wallet.Api/Application/Responses/Evenbet/EvenbetDebitResponse.cs
@@ -3,6 +3,6 @@
 using Base;
 
 public sealed record EvenbetDebitResponse(
-    decimal Balance,
+    int Balance,
     string Timestamp,
     string TransactionId) : EvenbetCommonResponseWithTransaction(Balance, Timestamp, TransactionId);
